requirements.txt
# RAG2.0 - Production Requirements
# Core system dependencies for the complete RAG2.0 system

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Azure services
azure-identity>=1.15.0
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
openai>=1.3.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0

# Logging and monitoring
structlog>=23.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0 